refactor(codegen): extract Phase-B method body emission into locus/method.rs (Round 4c)
1 fn lifted into LocusMethodBodies<'ctx>: lower_locus_method_bodies
(~848 lines). The Phase-B pass that emits actual IR bodies for
lifecycle methods (birth / run / dissolve / accept / drain), user
`fn` members, mode methods (bulk / harmonic / resolution), and the
closure-eval synthetics (__birth_closures / __tick_closures /
__duration_closures / etc.) declared in Phase A.

Surface bumps to pub(crate):
- Cx fields: current_user_fn_ret, loops, deferred_dissolves,
  current_method_scratch
- Cx methods: push_dissolve_frame, flush_dissolve_frame,
  flush_dissolve_frame_kind, lower_block, lower_closure_check,
  lower_fallible_locus_method_body, open_method_scratch,
  close_method_scratch
- Structs: LoopFrame

Imports BlockEnd / Scope / TimeStdlib so the synthesized
duration-closure last_fire seeding works.

codegen.rs: 32205 -> 31358 (-847 lines).
